Hendrik's changes

+++ flatland_server/include/flatland_server/layer.h
@@ -74,7 +74,7 @@ class Layer : public Entity {
     std::string path_;
     double resolution_;
     MapDescription() {}
-    MapDescription(std::string path, float resolution)
+    MapDescription(std::string path, double resolution)
         : path_(path), resolution_(resolution) {}
   };
