+++ flatland_server/src/layer.cpp
@@ -69,7 +69,7 @@ Layer::Layer(b2World *physics_world, CollisionFilterRegistry *cfr,
              const YAML::Node &properties)
     : Entity(physics_world, names[0]),
       names_(names),
-      map_description_{map_path, static_cast<float>(resolution)},
+      map_description_{map_path, resolution},
       cfr_(cfr),
       viz_name_("layer/" + names[0]) {
   body_ = new Body(physics_world_, this, name_, color, origin, b2_staticBody,
